api/requirements.txt
asyncio-pool==0.5.2
attrs==21.4.0
cachetools==4.2.4
cattrs==1.10.0
decorator==5.1.1
deprecation==2.1.0
fastapi-utils==0.2.1
fastapi==0.66.1
motor==2.5.1
pydantic==1.9.0
pymongo==3.12.3
retry==0.9.2
toolz==0.11.2
shotgun-api3 @ git+https://github.com/shotgunsoftware/python-api.git@v3.2.6
